fix(server-plugin): not relay on modifyRsbuildConfig
