+++ .changeset/light-turtles-hammer.md
@@ -1,0 +1,6 @@
+---
+'@rsbuild/webpack': patch
+'@rsbuild/core': patch
+---
+
+fix(server-plugin): not relay on modifyRsbuildConfig
