+++ packages/shared/src/types/plugin.ts
@@ -113,7 +113,6 @@ export type Plugins = {
   preloadOrPrefetch: PluginsFn;
   performance: PluginsFn;
   define: PluginsFn;
-  server: PluginsFn;
 };
 
 /**
